A lot of service updates, allowing for submission of ratings (closes #21), cache eviction upon rating submission (closes #23), custom index definition (closes #10), and updates to the readme (closes #25)

+++ src/main/groovy/io/durbs/movieratings/model/ViewableRating.groovy
@@ -9,6 +9,6 @@ class ViewableRating {
 
   String username
   String usersGivenName
-  Double rating
+  Integer rating
   String comment
 }
